tg bot refactoring
