+++ src/telegram/publisher/index.ts
@@ -1,7 +1,6 @@
 import TelegramBot from 'node-telegram-bot-api';
 import fs from 'fs';
 import axios from 'axios';
-import Readable from 'stream';
 import { massSendMessageThroughQueue, massSendPhotoThroughQueue, sendMessageWithSave, sendPhotoWithSave } from '../handlers/utils';
 import { getUserData } from '../../models/user';
 import { getAdminSession } from './session';
@@ -9,11 +8,9 @@ import { commands, menu } from './types';
 import { notABusyRegex } from '../../utils/text';
 import { adminCmdPreprocess, setupBotMenu } from './functions';
 import { Bot } from '../../telegram/bot';
-
-const api_token = process.env.TELEGRAM_PUBLISHER_API_TOKEN;
-const photoDirectory = "../../../downloads"
-
-export const publisherBot = api_token? new TelegramBot(api_token || '', { polling: true }) : null;
+import { photoDirectory, publisherBot, publisher_api_token } from './initial';
+import { mediaHandler } from './handlers/mediaHandler';
+import { textHandler } from './handlers/textHandler';
 
 export function initPublisherBot() {
     if (!fs.existsSync(photoDirectory)) {
@@ -23,16 +20,18 @@ export function initPublisherBot() {
   console.log('Publisher started');
   if (!publisherBot) return;
   setupBotMenu(publisherBot, menu);
-
+  if (!publisherBot) return;
 
 
   publisherBot.onText(/\/start/, async (msg) => {
+    if (!publisherBot) return;
     const chat = await adminCmdPreprocess(publisherBot, msg);
     if (!chat) return;
     sendMessageWithSave(publisherBot, chat, `Choose /newpost to start posting`);
     const session = getAdminSession(chat);
   });
   publisherBot.onText(/\/newpost/, async (msg) => {
+    if (!publisherBot) return;
     const chat = await adminCmdPreprocess(publisherBot, msg);
     if (!chat) return;
     const session = getAdminSession(chat);
@@ -43,6 +42,7 @@ export function initPublisherBot() {
     sendMessageWithSave(publisherBot, chat, `All right, a new post. Enter a post conent below:`);
   });
   publisherBot.onText(/\/addkeyboard/, async (msg) => {
+    if (!publisherBot) return;
     const chat = await adminCmdPreprocess(publisherBot, msg);
     if (!chat) return;
     const session = getAdminSession(chat);
@@ -54,6 +54,7 @@ export function initPublisherBot() {
     session.setLastAction("enter_keyboard");
   });
   publisherBot.onText(/\/confirmpost/, async (msg) => {
+    if (!publisherBot) return;
     const chat = await adminCmdPreprocess(publisherBot, msg);
     if (!chat) return;
     const session = getAdminSession(chat);
@@ -84,6 +85,7 @@ export function initPublisherBot() {
     }
   });
   publisherBot.onText(/\/cancelpost/, async (msg) => {
+    if (!publisherBot) return;
     const chat = await adminCmdPreprocess(publisherBot, msg);
     if (!chat) return;
     const session = getAdminSession(chat);
@@ -95,102 +97,11 @@ export function initPublisherBot() {
   });
 
   publisherBot.on('message', async (msg) => {
-    const chat = await adminCmdPreprocess(publisherBot, msg);
-    if (!chat) return;
-    if (!msg.text || !notABusyRegex(msg.text, commands)) {
-       return;
-    }
-    const session = getAdminSession(chat);
-    const action = session.getLastAction();
-    if (action === "init_post") {
-        session.textPost = msg.text;
-        console.log("Text: ", session.textPost);  
-        sendMessageWithSave(publisherBot, chat, `Look at your post and send it if ok: `);
-        setTimeout(() => {
-            sendMessageWithSave(publisherBot, chat, msg.text || "", {
-                parse_mode: "HTML"
-            });
-        }, 1101);
-        return;
-    }
-    if (action === "enter_keyboard") {
-        const chat = await adminCmdPreprocess(publisherBot, msg);
-        if (!chat) return;
-        const keyboardInfo = msg?.text.split(" ");
-        if (!keyboardInfo || keyboardInfo.length < 2) {
-            sendMessageWithSave(publisherBot, chat, `Invalid entry`);
-            return;
-        }
-        session.postKeyboard = [
-            [
-                {
-                    text: keyboardInfo[0], 
-                    url: keyboardInfo[1]
-                }
-            ]
-        ]
-        sendMessageWithSave(publisherBot, chat, `Look at your post and send it if ok: `);
-        setTimeout(() => {
-            sendMessageWithSave(publisherBot, chat, "", {
-                reply_markup: {
-                    inline_keyboard: session.postKeyboard
-                }
-            });
-        }, 1101);
-        return;
-    }
-    session.setLastAction("post_written");
+    await textHandler(msg)
   });
 
   publisherBot.on("photo", async (msg) => {
-    const chat = await adminCmdPreprocess(publisherBot, msg);
-    if (!chat) return;
-
-    const session = getAdminSession(chat);
-    const action = session.getLastAction();
-
-    if (action === "init_post") {
-        if (!msg.photo || msg.photo.length === 0) {
-            sendMessageWithSave(publisherBot, chat, "No photo in message");
-            return;
-        }
-        const photo = msg.photo[msg.photo.length - 1];  // Use the highest resolution photo
-        const file = await publisherBot.getFile(photo.file_id);
-        const fileUrl = `https://api.telegram.org/file/bot${api_token}/${file.file_path}`;
-
-        const response = await axios({
-            url: fileUrl,
-            method: 'GET',
-            responseType: 'stream'
-        });
-
-        const localFilePath = `${photoDirectory}/${file.file_path?.split('/').pop()}`;
-        console.log("loaded photo: ", localFilePath);
-        const writer = fs.createWriteStream(localFilePath);
-        /* const msg = await Bot.sendPhoto(chat, file.file_unique_id, {
-            caption: message,
-            ...options
-          }); */
-        response.data.pipe(writer);
-
-        await new Promise((resolve, reject) => {
-            writer.on('finish', () => resolve(localFilePath));
-            writer.on('error', reject);
-        });
-        const newFile = await sendPhotoWithSave(Bot, chat, localFilePath, "", true, {});
-        if (!newFile) {
-            sendMessageWithSave(publisherBot, chat, "Failed to resend photo");
-            return;
-        }
-        session.photoPost = { img: newFile || "", text: msg.caption};
-        sendMessageWithSave(publisherBot, chat, `Look at your photo post and send it if ok: `);
-        setTimeout(() => {
-            publisherBot.sendPhoto(chat, photo.file_id, {
-                caption: session.photoPost?.text,
-                parse_mode: "HTML"
-            });
-        }, 1101);
-        session.setLastAction("post_written");
-    }
+    await mediaHandler(msg)
   })
+
 }
